+++ packages/draw/src/geometry.component.ts
@@ -1,5 +1,5 @@
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, OnDestroy, OnInit, ViewContainerRef } from '@angular/core';
-import { PlaitBoard, PlaitPluginElementComponent, PlaitPluginElementContext, OnContextChanged } from '@plait/core';
+import { PlaitBoard, PlaitPluginElementComponent, PlaitPluginElementContext, OnContextChanged, updateForeignObject } from '@plait/core';
 import { Subject } from 'rxjs';
 import { GeometryShape, PlaitGeometry } from './interfaces/geometry';
 import { GeometryShapeGenerator } from './generator/geometry-shape.generator';
@@ -9,6 +9,8 @@ import { getTextRectangle } from './utils/geometry';
 import { ActiveGenerator, getRectangleByPoints } from '@plait/common';
 import { DefaultGeometryActiveStyle } from './constants/geometry';
 import { getStrokeWidthByElement } from './utils/geometry-style/stroke';
+import { PlaitText } from './interfaces';
+import { DefaultTextProperty } from './constants';
 
 @Component({
     selector: 'plait-draw-geometry',
@@ -40,20 +42,7 @@ export class GeometryComponent extends PlaitPluginElementComponent<PlaitGeometry
             }
         });
         this.shapeGenerator = new GeometryShapeGenerator(this.board);
-
-        this.textManage = new TextManage(
-            this.board,
-            this.viewContainerRef,
-            () => {
-                return getTextRectangle(this.element);
-            },
-            (textManageRef: TextManageRef) => {
-                const height = textManageRef.height;
-                if (textManageRef.newValue) {
-                    DrawTransform.setText(this.board, this.element, textManageRef.newValue, height);
-                }
-            }
-        );
+        this.initializeTextManage();
     }
 
     ngOnInit(): void {
@@ -91,8 +80,30 @@ export class GeometryComponent extends PlaitPluginElementComponent<PlaitGeometry
     updateText() {
         this.textManage.updateText(this.element.text);
         this.textManage.updateRectangle();
-        const textWidth = getTextRectangle(this.element).width;
-        this.textManage.updateWidth(textWidth);
+        if (!(this.element as PlaitText)?.autoSize) {
+            const textWidth = getTextRectangle(this.element).width;
+            this.textManage.updateWidth(textWidth);
+        }
+    }
+
+    initializeTextManage() {
+        const width = getTextRectangle(this.element).width;
+        this.textManage = new TextManage(this.board, this.viewContainerRef, {
+            getRectangle: () => {
+                return getTextRectangle(this.element);
+            },
+            onValueChangeHandle: (textManageRef: TextManageRef) => {
+                const height = textManageRef.height / this.board.viewport.zoom;
+                const width = textManageRef.width / this.board.viewport.zoom;
+
+                if (textManageRef.newValue) {
+                    DrawTransform.setText(this.board, this.element, textManageRef.newValue, width, height);
+                } else {
+                    DrawTransform.setTextSize(this.board, this.element, width, height);
+                }
+            },
+            maxWidth: (this.element as PlaitText)?.autoSize ? DefaultTextProperty.maxWidth : width
+        });
     }
 
     ngOnDestroy(): void {