+++ packages/draw/src/plugins/with-draw.ts
@@ -4,8 +4,9 @@ import { LineComponent } from '../line.component';
 import { PlaitDrawElement } from '../interfaces';
 import { getRectangleByPoints } from '@plait/common';
 import { withDrawHotkey } from './with-draw-hotkey';
-import { withGeometryCreate } from './with-geometry-create';
+import { withGeometryCreateByDraw, withGeometryCreateByDrag } from './with-geometry-create';
 import { withDrawFragment } from './with-draw-fragment';
+import { getTextRectangle } from '../utils';
 
 export const withDraw = (board: PlaitBoard) => {
     const { drawElement, getRectangle, isHitSelection, isMovable, dblclick } = board;
@@ -29,6 +30,15 @@ export const withDraw = (board: PlaitBoard) => {
     board.isHitSelection = (element: PlaitElement, range: Range) => {
         if (PlaitDrawElement.isGeometry(element)) {
             const client = getRectangleByPoints(element.points);
+
+            if (element.textHeight > client.height) {
+                const textClient = getTextRectangle(element);
+                return (
+                    RectangleClient.isHit(RectangleClient.toRectangleClient([range.anchor, range.focus]), client) ||
+                    RectangleClient.isHit(RectangleClient.toRectangleClient([range.anchor, range.focus]), textClient)
+                );
+            }
+
             return RectangleClient.isHit(RectangleClient.toRectangleClient([range.anchor, range.focus]), client);
         }
         return isHitSelection(element, range);
@@ -51,6 +61,5 @@ export const withDraw = (board: PlaitBoard) => {
         dblclick(event);
     };
 
-    return withGeometryCreate(withDrawFragment(withDrawHotkey(board)));
-
+    return withGeometryCreateByDrag(withGeometryCreateByDraw(withDrawFragment(withDrawHotkey(board))));
 };