+++ packages/draw/src/transforms/geometry-text.ts
@@ -1,14 +1,48 @@
-import { PlaitBoard, Transforms } from '@plait/core';
+import { PlaitBoard, Point, Transforms } from '@plait/core';
 import { Element } from 'slate';
-import { PlaitGeometry } from '../interfaces';
+import { PlaitGeometry, PlaitText } from '../interfaces';
+import { DefaultTextProperty, ShapeDefaultSpace } from '../constants';
 
-export const setText = (board: PlaitBoard, element: PlaitGeometry, text: Element, textHeight: number) => {
+const normalizePoints = (board: PlaitBoard, element: PlaitGeometry, width: number, textHeight: number) => {
+    let points = element.points;
+    let autoSize = (element as PlaitText).autoSize;
+    const defaultSpace = ShapeDefaultSpace.rectangleAndText;
+
+    if (autoSize) {
+        points = [points[0], [points[0][0] + width + defaultSpace * 2, points[0][1] + textHeight]];
+
+        if (width >= DefaultTextProperty.maxWidth) {
+            points = [points[0], [points[0][0] + width + defaultSpace * 2, points[0][1] + textHeight]];
+        }
+    }
+
+    return { points };
+};
+
+export const setText = (board: PlaitBoard, element: PlaitGeometry, text: Element, width: number, textHeight: number) => {
     const newElement = {
         text,
-        textHeight
+        textHeight,
+        ...normalizePoints(board, element, width, textHeight)
     };
 
     const path = board.children.findIndex(child => child === element);
 
     Transforms.setNode(board, newElement, [path]);
+};
+
+export const setTextSize = (board: PlaitBoard, element: PlaitGeometry, textWidth: number, textHeight: number) => {
+    if ((element as PlaitText).autoSize) {
+        const newElement = {
+            textHeight,
+            ...normalizePoints(board, element, textWidth, textHeight)
+        };
+        const isPointsEqual =
+            Point.isEquals(element.points[0], newElement.points[0]) && Point.isEquals(element.points[1], newElement.points[1]);
+        const isTextHeightEqual = Math.round(textHeight) === Math.round(element.textHeight);
+        if (!isPointsEqual || !isTextHeightEqual) {
+            const path = board.children.findIndex(child => child === element);
+            Transforms.setNode(board, newElement, [path]);
+        }
+    }
 };