+++ packages/draw/src/transforms/index.ts
@@ -1,8 +1,9 @@
 import { insertText, insertGeometry } from './geometry';
-import { setText } from './geometry-text';
+import { setText, setTextSize } from './geometry-text';
 
 export const DrawTransform = {
     setText,
     insertGeometry,
-    insertText
+    insertText,
+    setTextSize
 };