+++ packages/flow/src/draw/edge.ts
@@ -2,25 +2,26 @@ import { PlaitBoard, drawRoundRectangle, RectangleClient, drawArrow } from '@pla
 import { RoughSVG } from 'roughjs/bin/svg';
 import { getEdgePoints, getEdgeStyle } from '../utils/edge/edge';
 import { FlowEdge } from '../interfaces/edge';
+import { FlowRenderMode } from '../interfaces/flow';
 
-export const drawEdge = (board: PlaitBoard, roughSVG: RoughSVG, edge: FlowEdge, active = false) => {
+export const drawEdge = (board: PlaitBoard, roughSVG: RoughSVG, edge: FlowEdge, mode: FlowRenderMode) => {
     const [pathPoints] = getEdgePoints(board, edge);
-    const edgeStyles = getEdgeStyle(edge, active);
+    const edgeStyles = getEdgeStyle(edge, mode);
     return roughSVG.linearPath(
         pathPoints.map(item => [item.x, item.y]),
         edgeStyles
     );
 };
 
-export const drawEdgeLabel = (roughSVG: RoughSVG, edge: FlowEdge, textBackgroundRect: RectangleClient, active = false) => {
-    const edgeStyles = getEdgeStyle(edge, active);
+export const drawEdgeLabel = (roughSVG: RoughSVG, edge: FlowEdge, textBackgroundRect: RectangleClient, mode: FlowRenderMode) => {
+    const edgeStyles = getEdgeStyle(edge, mode);
     const { x, y, width, height } = textBackgroundRect;
     return drawRoundRectangle(roughSVG, x, y, x + width, y + height, edgeStyles, false, Math.min(width, height) / 2);
 };
 
-export const drawEdgeMarkers = (board: PlaitBoard, roughSVG: RoughSVG, edge: FlowEdge, active = false) => {
+export const drawEdgeMarkers = (board: PlaitBoard, roughSVG: RoughSVG, edge: FlowEdge, mode: FlowRenderMode) => {
     const [pathPoints] = getEdgePoints(board, edge);
-    const edgeStyles = getEdgeStyle(edge, active);
+    const edgeStyles = getEdgeStyle(edge, mode);
     const edgeMarkers: SVGGElement[] = [];
     if (edge.target.marker) {
         const [start, end] = pathPoints.splice(-2);