+++ packages/flow/src/flow-node.component.ts
@@ -2,7 +2,7 @@ import {
     ChangeDetectionStrategy,
     ChangeDetectorRef,
     Component,
-    ComponentRef,
+    NgZone,
     OnDestroy,
     OnInit,
     Renderer2,
@@ -15,6 +15,8 @@ import { drawNodeHandles } from './draw/handle';
 import { drawActiveMask, drawNode } from './draw/node';
 import { FlowNode } from './interfaces/node';
 import { FlowBaseData } from './interfaces/element';
+import { FlowRenderMode } from './interfaces/flow';
+import { setRelationEdgeSelected } from './utils/edge/relation-edge-selected';
 
 @Component({
     selector: 'plait-flow-node',
@@ -33,84 +35,105 @@ export class FlowNodeComponent<T extends FlowBaseData = FlowBaseData> extends Pl
 
     handlesG: SVGGElement | null = null;
 
-    constructor(public cdr: ChangeDetectorRef, public viewContainerRef: ViewContainerRef, public render2: Renderer2) {
+    activeG: SVGGElement | null = null;
+
+    constructor(
+        public cdr: ChangeDetectorRef,
+        public viewContainerRef: ViewContainerRef,
+        public render2: Renderer2,
+        public ngZone: NgZone
+    ) {
         super(cdr);
     }
 
     ngOnInit(): void {
         super.ngOnInit();
-        this.textManage = new TextManage(this.board, this.viewContainerRef, () => {
-            const { x, y } = normalizePoint(this.element.points![0]);
-            const width = this.element.width;
-            const height = this.element.height;
-            return { x, y, width, height };
+        this.textManage = new TextManage(this.board, this.viewContainerRef, {
+            getRectangle: () => {
+                const { x, y } = normalizePoint(this.element.points![0]);
+                const width = this.element.width;
+                const height = this.element.height;
+                return { x, y, width, height };
+            }
         });
         this.roughSVG = PlaitBoard.getRoughSVG(this.board);
         this.drawElement();
-        this.drawRichtext();
     }
 
     onContextChanged(value: PlaitPluginElementContext<FlowNode, PlaitBoard>, previous: PlaitPluginElementContext<FlowNode, PlaitBoard>) {
-        if (value.element !== previous.element && this.initialized) {
-            this.updateElement(value.element, value.selected);
+        if (this.initialized && (value.element !== previous.element || value.selected !== previous.selected)) {
+            this.drawElement(value.element, value.selected ? FlowRenderMode.active : FlowRenderMode.default);
         }
-        if (this.initialized) {
+        if (previous.selected !== value.selected) {
             if (value.selected) {
-                this.setActiveNodeToTop();
-                this.drawActiveMask();
-                this.drawHandles();
-            } else if (previous.selected) {
-                this.destroyActiveMask();
-                this.destroyHandles();
+                // setTimeout 解决当多个节点关联 edge 有交集时，先执行清空在执行选中操作
+                setTimeout(() => {
+                    setRelationEdgeSelected(this.board, this.element.id, value.selected);
+                }, 0);
+            } else {
+                setRelationEdgeSelected(this.board, this.element.id, value.selected);
             }
         }
     }
 
-    setActiveNodeToTop() {
-        const parentElement = this.g.parentElement;
-        this.g.remove();
-        parentElement?.append(this.g);
+    drawElement(element: FlowNode = this.element, mode: FlowRenderMode = FlowRenderMode.default) {
+        // 处理节点高亮当前为 selected 不绘制
+        if (this.selected && mode !== FlowRenderMode.active) {
+            return;
+        }
+        this.drawNode(element);
+        this.drawRichtext(element);
+        this.drawActiveMask(element, mode);
+        this.drawHandles(element, mode);
+
+        if (mode === FlowRenderMode.default) {
+            this.g.append(this.nodeG!);
+            this.g.append(this.textManage.g);
+            this.activeG && this.activeG.remove();
+        } else {
+            this.activeG = this.activeG || createG();
+            this.activeG?.append(this.nodeG!);
+            this.activeG?.append(this.textManage.g);
+            if (mode === FlowRenderMode.active) {
+                this.activeG?.prepend(this.activeMaskG!);
+            }
+            this.activeG?.append(this.handlesG!);
+            PlaitBoard.getElementActiveHost(this.board).append(this.activeG!);
+        }
     }
 
-    drawElement(element: FlowNode = this.element) {
+    drawNode(element: FlowNode = this.element) {
         this.destroyElement();
         this.nodeG = drawNode(this.roughSVG, element);
-        this.g.append(this.nodeG);
     }
 
-    drawActiveMask(element: FlowNode = this.element) {
+    drawActiveMask(element: FlowNode = this.element, mode: FlowRenderMode = FlowRenderMode.default) {
         this.destroyActiveMask();
-        this.activeMaskG = drawActiveMask(this.roughSVG, element);
-        this.g.prepend(this.activeMaskG);
+        if (mode === FlowRenderMode.active) {
+            this.activeMaskG = drawActiveMask(this.roughSVG, element);
+        }
     }
 
     drawRichtext(element: FlowNode = this.element) {
         this.destroyRichtext();
         if (element.data?.text) {
-            this.textManage.draw(element.data.text);
-            this.textManage.g.classList.add('flow-node-richtext');
-            this.g.append(this.textManage.g);
+            this.ngZone.run(() => {
+                this.textManage.draw(element.data?.text!);
+                this.textManage.g.classList.add('flow-node-richtext');
+            });
         }
     }
 
-    drawHandles(element: FlowNode = this.element) {
+    drawHandles(element: FlowNode = this.element, mode: FlowRenderMode = FlowRenderMode.default) {
         this.destroyHandles();
-        const handles = drawNodeHandles(this.roughSVG, element);
-        this.handlesG = createG();
-        handles.map(item => {
-            this.handlesG?.append(item);
-            this.render2.addClass(item, 'flow-handle');
-        });
-        this.handlesG?.setAttribute('stroke-linecap', 'round');
-        this.g.append(this.handlesG);
-    }
-
-    updateElement(element: FlowNode = this.element, isActive = false) {
-        this.drawElement(element);
-        this.drawRichtext(element);
-        if (isActive) {
-            this.drawActiveMask(element);
-            this.drawHandles(element);
+        if (mode !== FlowRenderMode.default) {
+            this.handlesG = createG();
+            const handles = drawNodeHandles(this.roughSVG, element);
+            handles.map(item => {
+                this.handlesG?.append(item);
+                this.render2.addClass(item, 'flow-handle');
+            });
+            this.handlesG?.setAttribute('stroke-linecap', 'round');
         }
     }
 
@@ -137,5 +160,15 @@ export class FlowNodeComponent<T extends FlowBaseData = FlowBaseData> extends Pl
 
     destroyRichtext() {
         this.textManage.destroy();
+    }
+
+    ngOnDestroy(): void {
+        super.ngOnDestroy();
+        this.destroyElement();
+        this.destroyHandles();
+        this.destroyActiveMask();
+        this.destroyRichtext();
+        this.activeG?.remove();
+        this.activeG = null;
     }
 }