+++ packages/flow/src/plugins/with-edge-create.ts
@@ -1,19 +1,5 @@
-import {
-    PlaitBoard,
-    PlaitPlugin,
-    isSelectedElement,
-    toPoint,
-    transformPoint,
-    drawLine,
-    idCreator,
-    throttleRAF,
-    removeSelectedElement,
-    getSelectedElements,
-    PlaitElement,
-    drawCircle
-} from '@plait/core';
+import { PlaitBoard, PlaitPlugin, toPoint, transformPoint, drawLine, idCreator, throttleRAF, drawCircle } from '@plait/core';
 import { FlowNode } from '../interfaces/node';
-import { FlowNodeComponent } from '../flow-node.component';
 import { FlowElementType } from '../interfaces/element';
 import { isEdgeDragging } from '../utils/edge/dragging-edge';
 import { destroyAllNodesHandle, drawAllNodesHandle } from '../utils/node/render-all-nodes-handle';
@@ -23,6 +9,7 @@ import { getHitHandleByNode, HitNodeHandle } from '../utils/handle/node';
 import { getHitNode } from '../utils/node/get-hit-node';
 import { DEFAULT_HANDLE_STYLES, HANDLE_DIAMETER } from '../constants/handle';
 import { getHoverHandleInfo } from '../utils/handle/hover-handle';
+import { addPlaceholderEdgeInfo, deletePlaceholderEdgeInfo } from '../utils/edge/placeholder-edge';
 
 export const withEdgeCreate: PlaitPlugin = (board: PlaitBoard) => {
     const { mousedown, globalMousemove, globalMouseup } = board;
@@ -36,12 +23,8 @@ export const withEdgeCreate: PlaitPlugin = (board: PlaitBoard) => {
 
     board.mousedown = event => {
         const point = transformPoint(board, toPoint(event.x, event.y, PlaitBoard.getHost(board)));
-        const selectElements = getSelectedElements(board);
         if (hoveredNode) {
             sourceFlowNodeHandle = getHitHandleByNode(hoveredNode, point);
-            selectElements.map(item => {
-                removeSelectedElement(board, item);
-            });
         }
         mousedown(event);
     };
@@ -63,6 +46,7 @@ export const withEdgeCreate: PlaitPlugin = (board: PlaitBoard) => {
                         const circleElement = drawCircle(PlaitBoard.getRoughSVG(board), point, HANDLE_DIAMETER, DEFAULT_HANDLE_STYLES);
                         circleElement?.setAttribute('stroke-linecap', 'round');
                         placeholderEdge.append(circleElement);
+                        addPlaceholderEdgeInfo(board, placeholderEdge);
                     }
                     PlaitBoard.getHost(board).append(placeholderEdge);
                     if (drawNodeHandles) {
@@ -105,20 +89,8 @@ export const withEdgeCreate: PlaitPlugin = (board: PlaitBoard) => {
                     if (newHitNode == hoveredNode || isHitHoveredNodeHandle) {
                         return;
                     }
-                    const isSelectedHoveredNode = isSelectedElement(board, hoveredNode);
-                    if (!isSelectedHoveredNode) {
-                        // destroy handles
-                        const flowNodeComponent = PlaitElement.getComponent(hoveredNode) as FlowNodeComponent;
-                        flowNodeComponent?.destroyHandles();
-                    }
                 }
-
                 hoveredNode = newHitNode;
-                if (hoveredNode) {
-                    // draw handles
-                    const flowNodeComponent = PlaitElement.getComponent(hoveredNode) as FlowNodeComponent;
-                    flowNodeComponent?.drawHandles(hoveredNode);
-                }
             }
         }
         globalMousemove(event);
@@ -130,6 +102,7 @@ export const withEdgeCreate: PlaitPlugin = (board: PlaitBoard) => {
             sourceFlowNodeHandle = null;
             targetFlowNodeHandle = null;
             placeholderEdge?.remove();
+            deletePlaceholderEdgeInfo(board);
             deleteCreateEdgeInfo(board);
             drawNodeHandles = true;
             destroyAllNodesHandle(board, flowNodeElements);