+++ packages/flow/src/utils/index.ts
@@ -6,6 +6,9 @@ export * from './edge/get-smooth-step-edge';
 export * from './edge/is-hit-edge';
 export * from './edge/text';
 export * from './edge/label-space';
+export * from './edge/get-hit-edge';
+export * from './edge/placeholder-edge';
+export * from './edge/relation-edge-selected';
 
 export * from './handle/get-default-handles';
 export * from './handle/get-handle-position';