+++ packages/flow/src/utils/node/render-all-nodes-handle.ts
@@ -3,12 +3,14 @@ import { FlowNode } from '../../interfaces/node';
 import { PlaitBoard, PlaitElement, isSelectedElement } from '@plait/core';
 import { getFlowElementsByType } from './get-node';
 import { FlowNodeComponent } from '../../flow-node.component';
+import { FlowRenderMode } from '../../interfaces/flow';
 
 export function drawAllNodesHandle(board: PlaitBoard) {
     const flowNodeElements = getFlowElementsByType(board, FlowElementType.node) as FlowNode[];
     flowNodeElements.map(item => {
         const flowNodeComponent = PlaitElement.getComponent(item) as FlowNodeComponent;
-        flowNodeComponent.drawHandles();
+        flowNodeComponent.drawHandles(item, FlowRenderMode.active);
+        PlaitBoard.getElementActiveHost(board).append(flowNodeComponent.handlesG!);
     });
     return flowNodeElements;
 }
@@ -16,9 +18,9 @@ export function drawAllNodesHandle(board: PlaitBoard) {
 export function destroyAllNodesHandle(board: PlaitBoard, flowNodeElements: FlowNode[]) {
     flowNodeElements.map(item => {
         const flowNodeComponent = PlaitElement.getComponent(item) as FlowNodeComponent;
-        flowNodeComponent.destroyHandles();
+        !flowNodeComponent.selected && flowNodeComponent.destroyHandles();
         if (isSelectedElement(board, item)) {
-            flowNodeComponent.destroyActiveMask();
+            !flowNodeComponent.selected && flowNodeComponent.destroyActiveMask();
         }
     });
 }